Maximum number of elements 10000 -> 100000

+++ src/sorting/InformationSort.java
@@ -1,23 +1,23 @@
 package sorting;
 
 public class InformationSort {
-    private final int comparisons;
-    private final int swaps;
+    private final long comparisons;
+    private final long swaps;
     private final long time;
     private final long memory;
 
-    public InformationSort(int comparisons, int swaps, long time, long memory) {
+    public InformationSort(long comparisons, long swaps, long time, long memory) {
         this.comparisons = comparisons;
         this.swaps = swaps;
         this.time = time;
         this.memory = memory;
     }
 
-    public int getComparisons() {
+    public long getComparisons() {
         return comparisons;
     }
 
-    public int getSwaps() {
+    public long getSwaps() {
         return swaps;
     }
